Create hot key bindings for instant switches
Created the new key bindings and added properties in
DependencyInjector which get the default text manager and the
instant text manager. The App class uses those according to the
hot key pressed. When the text is switched via settings window
the default text manager is used.

The InstantClipboardTextManager now accesses the text in the
clipboard only if it contains any text.

Also added some minor fixes not worth mentioning.

As before, instant switching doesn't work yet.