Implement starting the app on system startup
Add the ability to start the app when the user logs in by adding a
regsitry value. Also, add the ability to manually enable/disable it in
the settings app. Set it to run at startup after installation.

Make both the service app and the settings app use their location as the
working directory regardless of whether they were launched in it or not.
This is kind of a hack, but it's needed becuase when the service app is
started at system startup, its working directory is different, and it
needs to read the appsettings.json file in its folder.

Also, start the settings app after installation through the File
Explorer instead of directly, because the installer runs as an elevated
process and if the settings app is run directly by it, it will be
elevated as well. This is not great, because the app adds a value to the
HKEY_CURRENT_USER registry hive, and it cannot do that if it's elevated.
This solution is a hack as well, but much easier to implement compared
to alternatives.