Use another hack during the app installation
Previously the settings app configured the service to run at system
startup if if didn't find the settings file, which is not ideal, becuase
this file can be retained between installations.

Instead, the installer now creates an empty file before launching the
settings app for the first time. The app checks whether this file
exists, and if yes, then it configures the service to start at system
startup, and deletes this file. This is a hack as I couldn't think of
better ways to handle this (becuase the app is started through the File
Explorer the installer cannot pass any command-line arguments or
environment variables).