Upgrade to .NET 5
Upgrade from .NET Core 3.1 to .NET 5. Change the target of
KeyboardSwitch.Windows to be Windows-only and thus introduce
preprocessor constants to KeyboardSwitch and KeyboardSwitch.Settings to
keep them cross-platform.

The app itself can be built at this commit, but it's untested (not even
started) and probably not going to run correctly, because the libraries
are still at old versions which may not support .NET 5.

 #32