Add a minor hack to the keyboard hook
Sometimes libuiohook doesn't appear to send the 'key released' event and
it seems that it doesn't do that when the app simulates pressing Ctrl+C
or Ctrl+V in instant mode. Because of that, all following switches will
not occur since there's a key that's prepetually 'pressed'. Add a check
which evicts all 'pressed' keys if the last key press was more than 3
seconds ago. This is obviously not an ideal solution, but at least it
will not require the user to restart the service to continue using the
app.