Make the clipboard service cross-platform
Use the TextCopy library which is cross-platform instead of directly
interfacing with the clipboard. This library uses bash and xsel on Linux
so now the app is dependent on those there.

This makes the KeyboardSwitch app minimally viable on the Linux
platform. A lot of things should be implemented to make it usable on the
same level as on Windows, but at least it works!

Also, fix a bug with the X11 layout service.

 #44