Rename KeyboardSwitch.Common to KeyboardSwitch.Core
"Common" just doesn't cut it anymore. I thought in the beginning that
this project is going to contain basically just the interfaces and that
their implementations are going to reside in platform-specific projects.
Turns out that most functionality, and most importantly, the main logic,
is cross-platform, so it's contained in this project. As such, it
doesn't just contain the common functionality - it contains the core of
the app.