Fix a bug with char mapping text boxes
Because of an Avalonia bug, using shared size groups with text boxes is
basically impossible because the UI freezes when the text box tries to
resize itself. Remove shared size groups. This means that the UX of the
char mapping is destroyed, because it's much more difficult to determine
which character corresponds to other characters in different layouts.
Alternative solutions to this problem should be considered. Working
badly is better than not working at all.