Make layout names in char mapping view the same width
This change is the fix of the previous change - the layout names can in
fact be included in the shared size group - it's only the text boxes
that shouldn't be included. This means that the text boxes won't be the
same width when they have a different number of characters, but that
shouldn't be a thing anyway.