Implement simulating copying and pasting in Linux
Now the instant switcing mode works correctly on Linux.

The simulation is implemented using X11's XTest extension, which has to
be installed separately and this means that the app now depends on one
more thing (the libxtst-dev on Ubuntu).

The X11-specific keysyms are specified only for left control, C, and V
for now. When the auto-configuration feature will be implemented, more
keysyms will be added and probably moved into an enum.

 #48