Implement switching layouts via keyboard simulation
Since switching layouts directly via X11 doesn't seem to work and
viable alternatives haven't been found, it is now implemented through
simulating pressing of keys which the user uses to switch layouts. These
key combinations need to be configurable since users can change them in
the system.

Specifing which keys should be used for switching is implemented as a
custom text box which tracks various key presses (not all, but the most
common ones are covered). Pressing the super keys on Ubuntu 20.04 is not
captured, so buttons to add them manually were added as well.

This is only needed for Linux for now, and will work on macOS unless
switching layouts directly is implemented there as well.

 #43