Remove automatic retrying when starting the service
Retrying within the app didn't prove to be useful. It was added as a
hack so that the app can start more or less normally when started by
systemd at login. That doesn't work since X11 simply kills the process
if the connection cannot be established. Also, systemd integration is
dropped altogether, so this feature is not needed anymore.

 #57