Improve the CLI experience
Add a help message when 'help' is specified. Run the main app flow only
when there are no arguments. Add checking whether the app is running
from the terminal. Also, add exit codes.

All of this works only on Linux though - on Windows the servie app is
specified as a GUI app (even though it has no UI) so that the console
window is not shown. And GUI apps cannot output anything to the console.

 #58