Add a hack so that dpkg doesn't remove /opt
When dpkg uninstalls a package, it tries to remove all directories which
contained that package's files (so, it will delete /opt/keyboard-switch
and try to delete /opt). If /opt doesn't contain other files, it will be
deleted which is not desirable. Add a post-remove file which recreates
/opt if it was deleted. This is a hack since a better way would be to
not let dpkg remove /opt in the first place.