Fix the projects' setup and configuration
Turns out that the app is not built correctly when using net6.0-macos.
Use OS information instead of TFM to determine which projects the apps
should reference. This also means that KeyboardSwitch.Linux cannot be
used as a simple 'else' branch and a preprocessor constant should be
defined for it as well.

 #60