Refactor the hook service to work with a blocking hook
Since on macOS the global hook apparantly should run on the main thread
(at least, that's how it is with SharpHook), the hook service had to be
refactored a little bit. This doesn't make it work on macOS yet.

Also, add a blocking reactive global hook - this class will most
probably be added to SharpHook in the nearest future.

 #60