Refactor appsettings.json
Remove the generic appsettings.json file as it has been empty for a long
time. Instead, always use platform-specific application.env.json files
and rename them to application.json when building installers.

Also, determine whether the settings app is inside of a macOS bundle. if
it is, then the settings will not be in the same directory, but in the
resources directory of the bundle. Lastly, set the service path to
/opt/keyboard-switch/KeyboardSwitch on macOS since the plan is that the
service app and the settings app won't be in the same directory on that
system.