Implement the macOS installer for version 4.2.0
This required changing the app bundle ID for the service app since macOS
really doesn't like changing the location of an app bundle during a
package installation, so it appears to be basically impossible.

Also, the installer now removes the old version of Keyboard Switch
before installing the new one.

GH-92