Add a hack which fixes switching layouts on Linux
For some inexplicable reason switching the current keyboard layout stops
working if xsel integration is enabled. It would work if the main loop
was kept running, but this causes other problems on several Linux
distributions. The only way I've found to fix that is to set the X11
synchronous mode on. It doesn't appear to break anything (both the xsel
clipboard integration and native clipboard integration seem to work fine
with it), so let's roll with it.